split core function from menu functions, restore old heat & preheat menu
