+++ TFT/src/User/Menu/FeatureSettings.c
@@ -90,10 +90,10 @@ LISTITEM settingPage[SKEY_COUNT] = {
   {ICONCHAR_TOGGLE_ON,  LIST_TOGGLE,        LABEL_INVERT_YAXIS,             LABEL_BACKGROUND  },
   {ICONCHAR_TOGGLE_ON,  LIST_TOGGLE,        LABEL_INVERT_ZAXIS,             LABEL_BACKGROUND  },
   #ifdef PS_ON_PIN
-    {ICONCHAR_TOGGLE_ON,  LIST_TOGGLE,        LABEL_AUTO_SHUT_DOWN,           LABEL_BACKGROUND  },
+    {ICONCHAR_TOGGLE_ON,  LIST_TOGGLE,      LABEL_AUTO_SHUT_DOWN,           LABEL_BACKGROUND  },
   #endif
   #ifdef FIL_RUNOUT_PIN
-    {ICONCHAR_BLANK,      LIST_CUSTOMVALUE,   LABEL_FILAMENT_SENSOR,          LABEL_OFF         },
+    {ICONCHAR_BLANK,     LIST_CUSTOMVALUE,   LABEL_FILAMENT_SENSOR,          LABEL_OFF         },
   #endif
   {ICONCHAR_BLANK,      LIST_CUSTOMVALUE,   LABEL_MOVE_SPEED,               LABEL_NORMAL_SPEED},
   {ICONCHAR_TOGGLE_ON,  LIST_TOGGLE,        LABEL_SEND_START_GCODE,         LABEL_BACKGROUND  },
@@ -113,7 +113,7 @@ LISTITEM settingPage[SKEY_COUNT] = {
   {ICONCHAR_BLANK,      LIST_TOGGLE,        LABEL_ST7920_FULLSCREEN,        LABEL_OFF         },
   #endif
   // Keep reset settings always at the bottom of the settings menu list.
-  {ICONCHAR_BLANK,      LIST_MOREBUTTON,    LABEL_SETTING_RESET,           LABEL_BACKGROUND }
+  {ICONCHAR_BLANK,      LIST_MOREBUTTON,    LABEL_SETTING_RESET,            LABEL_BACKGROUND  }
 };
 
 void resetSettings(void)
