+++ TFT/src/User/Menu/SettingsMenu.h
@@ -1,0 +1,11 @@
+#ifndef _SETTINGSMENU_H_
+#define _SETTINGSMENU_H_
+
+#include "stdint.h"
+#include "Settings.h"
+
+extern const u32 item_baudrate[ITEM_BAUDRATE_NUM];
+
+void menuSettings(void);
+
+#endif
