+++ TFT/src/User/includes.h
@@ -54,6 +54,9 @@
 #include "flashStore.h"
 #include "parseACK.h"
 #include "Selectmode.h"
+#include "Temperature.h"
+#include "Settings.h"
+#include "Printing.h"
 
 #include "extend.h"
 #include "list_item.h"
@@ -64,6 +67,7 @@
 #include "menu.h"
 #include "MainPage.h"
 #include "Heat.h"
+#include "PreheatMenu.h"
 #include "Move.h"
 #include "Home.h"
 #include "Print.h"
@@ -76,7 +80,8 @@
 
 #include "Extrude.h"
 #include "Fan.h"
-#include "Settings.h"
+#include "SettingsMenu.h"
+#include "PrintingMenu.h"
 #include "ScreenSettings.h"
 #include "MachineSettings.h"
 #include "FeatureSettings.h"
